css hack improves external links inside SVGs
TODO: We should implement the use of CSS ::after {content:url(...)} in
the wiki-client and then remove this hack.

Currently external links are expanded with an appended <img> tag.
<img> is not valid SVG. Chrome resolves the conflict by pretending the
<img> comes after the close tag of the SVG. That just looks funny.

The ::after pseudo-selector only works in the HTML. So the external
links inside an SVG do not get the icon added. But neither do they
leave an outlier <img> on display.

We think this compromise is preferable.